@@ -1,7 +1,7 @@
 [package]
 name = "clash-verge-service"
 version = "1.1.0"
-edition = "2021"
+edition = "2024"
 authors = ["zzzgydi", "Clash Verge Rev Team", "oomeow"]
 license = "GPL-3.0"
 repository = "https://github.com/oomeow/clash-verge-service.git"
